Abstract MobileNav from Header
