+++ src/components/_modules/header/index.js
@@ -3,19 +3,18 @@ import { useRouter } from 'next/router';
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { breakpoint } from 'src/settings';
-import { stripTags } from 'src/utils/strings';
 import { Date as DateComponent } from 'src/components';
-import { links } from './links';
+
 import { StyledHeader, HeroContainer, Nav, NavImage, NavItem } from './styles';
 import {
+  MobileNav,
   MobileNavItem,
   MobileNavMenu,
   MobileNavMenuWrapper,
   MobileNavOverlay,
   MobileNavWrapper,
   MobileNavListItem,
-} from './mobile-overlay';
+} from '../mobile-nav';
 
 const Glow = styled.div`
   --size: 1200px;
@@ -35,9 +34,6 @@ export const Header = (props) => {
 
   const { date, headline = null } = props;
   const [isVisible, setVisibility] = useState(false);
-  const [overlayTransitioned, setOverlayTransitioned] = useState(false);
-  const [showSocial, setShowSocial] = useState(false);
-  const router = useRouter();
 
   // TODO: Create util
   const formattedDate = date
@@ -49,126 +45,46 @@ export const Header = (props) => {
         .replace('.', '')
     : null;
 
-  useEffect(() => {
-    // Lock scroll when Nav overlay is visible
-    document.body.style.overflow = isVisible ? 'hidden' : 'scroll';
-
-    // Ensure Nav is hidden when viewport is larger than sm breakpoint
-    function handleResize() {
-      if (window.innerWidth > parseInt(breakpoint.sm)) {
-        setVisibility(false);
-        setOverlayTransitioned(false);
-      }
-    }
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, [isVisible]);
-
   return (
     <>
-      <MobileNavOverlay
-        $isVisible={isVisible}
-        onTransitionEnd={() => {
-          setOverlayTransitioned(isVisible ? true : false);
-          setShowSocial(isVisible);
-        }}
-      >
-        <MobileNavWrapper>
-          {links.map(({ to, label }, index) => (
-            <Link href={to} key={to} passHref>
-              <MobileNavItem
-                $isActive={router.pathname === to ? 'is-active' : ''}
-                $isVisible={overlayTransitioned}
-                onAnimationStart={() =>
-                  setShowSocial(index === links.length - 1)
-                }
-                style={{ animationDelay: `calc(${100 * index}ms)` }}
-              >
-                {label}
-              </MobileNavItem>
-            </Link>
-          ))}
-
-          <ul>
-            {allMdx &&
-              allMdx.edges.map(({ node }, index) => (
-                <MobileNavListItem
-                  key={node.fields.slug}
-                  $isVisible={showSocial}
-                  style={{
-                    animationDelay: `calc(${100 * (index + 1)}ms)`,
-                    textDecoration: 'underline',
-                  }}
-                >
-                  <Link key={node.frontmatter.title} href={node.fields.slug}>
-                    {stripTags(node.frontmatter.title)}
-                  </Link>
-                </MobileNavListItem>
-              ))}
-          </ul>
-
-          <ul>
-            {['Twitter', 'Instagram', 'GitHub'].map((link, index) => (
-              <MobileNavListItem
-                $isVisible={showSocial}
-                key={link}
-                style={{
-                  animationDelay: `calc(${100 * (index + 1)}ms)`,
-                }}
-              >
-                <a
-                  href={`https://${link.toLowerCase()}.com/bradcerasani`}
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  {link}
-                </a>
-              </MobileNavListItem>
-            ))}
-          </ul>
-        </MobileNavWrapper>
-      </MobileNavOverlay>
-
       <StyledHeader
         itemscope="itemscope"
         itemtype="https://schema.org/SiteNavigationElement"
       >
-        <MobileNavMenuWrapper>
-          <Link href="/" passHref>
-            <div
+        <Link href="/" passHref>
+          <div
+            style={{
+              marginBottom: 'var(--spaceMedium)',
+              position: 'relative',
+            }}
+          >
+            <Glow />
+            <h1
               style={{
-                marginBottom: 'var(--spaceMedium)',
-                position: 'relative',
+                lineHeight: '1.5',
+                fontFamily: 'var(--fontFamilySansSerif)',
+                fontSize: 'var(--fontSizeSmall)',
+                fontWeight: 500,
               }}
             >
-              <Glow />
-              <h1
+              Brad Cerasani
+              <span
                 style={{
-                  lineHeight: '1.5',
-                  fontFamily: 'var(--fontFamilySansSerif)',
-                  fontSize: 'var(--fontSizeSmall)',
-                  fontWeight: 500,
+                  color: 'var(--colorGreyLight)',
+                  display: 'block',
+                  fontWeight: 400,
                 }}
               >
-                Brad Cerasani
-                <span
-                  style={{
-                    color: 'var(--colorGreyLight)',
-                    display: 'block',
-                    fontWeight: 400,
-                  }}
-                >
-                  Design &amp; Engineering
-                </span>
-              </h1>
-            </div>
-          </Link>
-          <MobileNavMenu
-            onClick={() => setVisibility(!isVisible)}
-            $isActive={isVisible}
-          />
-        </MobileNavMenuWrapper>
+                Design &amp; Engineering
+              </span>
+            </h1>
+          </div>
+        </Link>
+
+        <MobileNav
+          onClick={() => setVisibility(!isVisible)}
+          $isActive={isVisible}
+        />
 
         {headline && (
           <HeroContainer>
