--- src/components/_modules/header/links.js
@@ -1,14 +1,0 @@
-export const links = [
-  {
-    to: '/',
-    label: 'Index',
-    src: '/images/nav/index@2x.webp',
-    alt: 'Homepage',
-  },
-  {
-    to: '/about',
-    label: 'About',
-    src: '/images/nav/about-80s.jpg',
-    alt: 'About Brad Cerasani',
-  },
-];
